VocalXpert An Intelligent Assistant/requirements.txt
# VocalXpert AI Assistant - Requirements

# ===== GUI Framework =====
PySide6>=6.5.0

# Legacy Tkinter support is built into Python

# ===== Speech Recognition & Text-to-Speech =====
SpeechRecognition>=3.10.0
pyttsx3>=2.90
playsound==1.2.2
PyAudio>=0.2.13
pocketsphinx>=5.0.0
vosk>=0.3.44

# ===== Computer Vision & Face Recognition =====
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
numpy>=1.20.0

# ===== Image Processing =====
Pillow>=9.0.0
pyscreenshot>=3.1
pynput>=1.7.6
pyautogui>=0.9.53

# ===== System & OS =====
psutil>=5.9.0
wmi; sys_platform == 'win32'

# ===== Web Scraping & Automation =====
beautifulsoup4>=4.11.0
requests>=2.28.0
wikipedia>=1.4.0
youtube-search-python>=1.6.2
selenium>=4.0.0
fake-useragent>=1.0.0
lxml>=4.9.0
aiohttp>=3.8.0

# ===== Proxy Management (Optional - enhances uncensored access) =====
free-proxy>=1.1.0
requests-ip-rotator>=1.0.0
proxyscrape>=0.3.0
PySocks>=1.7.1

# ===== Geolocation =====
geopy>=2.4.0

# ===== Translation =====
googletrans>=4.0.0

# ===== Environment Variables =====
python-dotenv>=1.0.0
